pyproject.toml
[project]
name = "TinyCTA"
version = "0.10.0"
description = "..."
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
authors = [{name = "Thomas Schmelzer", email = "thomas.schmelzer@gmail.com"}]
requires-python = ">=3.11"
dependencies = [
    "loguru>=0.7.3",
    "numpy>=2.0.0",
    "optuna>=4.8.0",
    "pandas>=2.2.3",
    "polars>=1.0.0",
    "pyarrow>=23.0.0",
    "pydantic>=2.13.3",
    "pyyaml>=6.0.3",
]
[project.urls]
repository = "https://github.com/tschm/tinycta"

[dependency-groups]
dev = [
    "pre-commit==4.5.1",
    "marimo==0.23.2",
    "plotly>=6.5",
    "pandas-stubs>=2.2"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tinycta"]

[tool.deptry.per_rule_ignores]
DEP002 = [
    "pyarrow",       # polars parquet backend
]

[tool.deptry.package_module_name_map]
numpy = "numpy"
pandas = "pandas"
pandas-stubs = []
pre-commit = "pre_commit"
marimo = "marimo"
plotly = "plotly"
pyyaml = "yaml"

